Refactor `types/index.ts`
Splits up `index.ts` into various files. `index` is now a barrel file.
