+++ src/lib/types/base.ts
@@ -1,0 +1,11 @@
+export interface WithId {
+	id: number;
+}
+
+export interface CreatedAt {
+	createdAt: Date;
+}
+
+export type Stored<T> = T & WithId;
+
+export type StoredId = WithId['id'];
