+++ src/lib/types/repository.ts
@@ -1,0 +1,9 @@
+import type { WithId, StoredId } from './base';
+
+export interface Repository<TInput, TOutput extends WithId> {
+	add(entry: TInput): Promise<StoredId>;
+	update(id: StoredId, patch: Partial<TInput>): Promise<void>;
+	remove(id: StoredId): Promise<void>;
+	getById(id: StoredId): Promise<TOutput | undefined>;
+	getAll(): Promise<TOutput[]>;
+}
