+++ src/lib/types/settings.ts
@@ -1,0 +1,6 @@
+import type { WithId } from './base';
+
+export interface AppSettings extends WithId {
+	key: string;
+	value: unknown;
+}
